Move import logic to JayveeImportResolver
