+++ libs/language-server/src/lib/services/import-resolver.ts
@@ -3,21 +3,41 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 import {
+  type AstNode,
   AstUtils,
+  DocumentCache,
   type LangiumDocument,
   type LangiumDocuments,
-  type URI,
+  URI,
   UriUtils,
 } from 'langium';
 
+import { type ExportDetails, getExportedElements } from '../ast';
 import {
+  type ExportableElement,
   type ImportDefinition,
   type JayveeModel,
   isJayveeModel,
 } from '../ast/generated/ast';
+import { getStdLib } from '../builtin-library/stdlib';
+import { type JayveeServices } from '../jayvee-module';
+
+interface ImportDetails {
+  element: ExportableElement;
+  importName: string;
+}
 
 export class JayveeImportResolver {
-  constructor(protected documents: LangiumDocuments) {}
+  protected readonly documents: LangiumDocuments;
+  protected readonly availableElementsPerDocumentCache: DocumentCache<
+    string,
+    ExportDetails[]
+  >; // DocumentCache becomes invalidated as soon the corresponding document is updated
+
+  constructor(services: JayveeServices) {
+    this.documents = services.shared.workspace.LangiumDocuments;
+    this.availableElementsPerDocumentCache = new DocumentCache(services.shared);
+  }
 
   resolveImport(importDefinition: ImportDefinition): JayveeModel | undefined {
     const resolvedDocument = this.resolveImportedDocument(importDefinition);
@@ -62,5 +82,116 @@ export class JayveeImportResolver {
     const modelPath = importDefinition.path;
 
     return UriUtils.resolvePath(dirUri, modelPath);
+  }
+
+  getImportedElements(model: JayveeModel): ImportDetails[] {
+    const importedElements: ImportDetails[] = [];
+    for (const importDefinition of model.imports) {
+      const importedDocument = this.resolveImportedDocument(importDefinition);
+      if (importedDocument === undefined) {
+        continue;
+      }
+
+      importedElements.push(
+        ...this.getImportedElementsFromDocument(
+          importDefinition,
+          importedDocument,
+        ),
+      );
+    }
+    return importedElements;
+  }
+
+  getImportedElementsFromDocument(
+    importDefinition: ImportDefinition,
+    importedDocument: LangiumDocument,
+  ): ImportDetails[] {
+    const publishedElements =
+      this.getPublishedElementsFromDocument(importedDocument);
+
+    if (importDefinition.useAll) {
+      return publishedElements.map((exportDetails) => {
+        return {
+          element: exportDetails.element,
+          importName: exportDetails.exportName,
+        };
+      });
+    }
+
+    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+    const namedImports = importDefinition.usedElements ?? [];
+
+    const importedElements: ImportDetails[] = [];
+    for (const namedImport of namedImports) {
+      const matchingExportedElement = publishedElements.find(
+        (x) => x.exportName === namedImport.element,
+      );
+      if (
+        matchingExportedElement === undefined ||
+        // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+        matchingExportedElement.element === undefined
+      ) {
+        continue;
+      }
+
+      const importedElementName = namedImport.alias ?? namedImport.element;
+      importedElements.push({
+        element: matchingExportedElement.element,
+        importName: importedElementName,
+      });
+    }
+    return importedElements;
+  }
+
+  getBuiltinElements(): ExportDetails[] {
+    const builtinUris = this.getBuiltinUris();
+
+    const importedDocuments = [...builtinUris].map((importedUri) =>
+      this.documents.getDocument(URI.parse(importedUri)),
+    );
+
+    const importedElements: ExportDetails[] = [];
+    for (const importedDocument of importedDocuments) {
+      if (importedDocument === undefined) {
+        continue;
+      }
+
+      importedElements.push(
+        ...this.getPublishedElementsFromDocument(importedDocument),
+      );
+    }
+
+    return importedElements;
+  }
+
+  getPublishedElementsFromDocument(
+    document: LangiumDocument<AstNode>,
+  ): ExportDetails[] {
+    const model = document.parseResult.value as JayveeModel;
+
+    const publishedElements = this.availableElementsPerDocumentCache.get(
+      document.uri,
+      'exports', // we only need one key here as it is on document basis
+      () => getExportedElements(model),
+    );
+    return publishedElements;
+  }
+
+  /**
+   * Gets all builtins' URIs
+   */
+  getBuiltinUris(): Set<string> {
+    const importedUris: Set<string> = new Set();
+
+    const builtins = getStdLib();
+    const uris = Object.keys(builtins);
+
+    for (const uri of uris) {
+      // without formatting, the document might not be found
+      const formattedUri = URI.parse(uri).toString();
+      importedUris.add(formattedUri);
+    }
+
+    return importedUris;
   }
 }
